Move the `Noop` test to `noop.rs`
